re_arrange two func to core.depmodel \    to avoid circul import
